blufor_headgear - reorganize into separate configs
